+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - Remove unnecessary limitation on `pcommon.Value.Equal` that slices have only primitive values. (#5865)
 - Add support to handle 404, 405 http error code as permanent errors in OTLP exporter (#5827)
 - Enforce scheme name restrictions to all `confmap.Provider` implementations. (#5861)
+- Add httpprovider to allow loading config files stored in HTTP (#5810)
 
 ### 🧰 Bug fixes 🧰